v0.40

+++ WpfApplication1/Classes/Files.cs
@@ -7,9 +7,16 @@ namespace WpfApplication1
     {
         public static void SaveToArchive(string curId,string text,string type)
         {
-            StreamWriter w = new StreamWriter(File.Create(@"" + Properties.Settings.Default.SaveArchive + "\\" + type + curId + ".txt"), Encoding.GetEncoding(1251));
+            if(Directory.Exists(@"" + Properties.Settings.Default.SaveArchive))
+            {
+                StreamWriter w = new StreamWriter(File.Create(@"" + Properties.Settings.Default.SaveArchive + "\\" + type + curId + ".txt"), Encoding.GetEncoding(1251));
                 w.WriteLine(text);
                 w.Close();
+            }
+            else
+            {
+                Directory.CreateDirectory(@"" + Properties.Settings.Default.SaveArchive);
+            }
         }
     }
 }
